refactor(internal/dao): 合并集合相关的 DAO 函数
- 将 InsertCollectionUser、DeleteCollectionUser、InsertCollectionProblem 和 DeleteCollectionProblem 函数合并到 collection.go 文件中
- 删除 collection_user.go 和 collection_problem.go 文件
- 通过合并相关功能简化代码结构，提高可维护性
